+++ CHANGELOG.md
@@ -57,3 +57,4 @@
 - Actual start and end dates must not be in the future
 - Activity XML includes 'iati-identifier' which is includes the reporting organisation
 - Activity XML includes 'iati-identifier' with an identifier that consists of all present levels of hierarchy: fund and/or programme
+- BEIS users (administrators) can mark other users as active or inactive